[wallet-ext] add account badge to the account selector dropdown (#9719)
## Description 
This PR adds a badge to the account selector dropdown so folks can
easily identify their accounts. I also made some small accessibility
improvements like using an unordered list and using a `button` vs a
clickable div.

<img width="307" alt="image"
src="https://user-images.githubusercontent.com/7453188/227027566-2e779fc9-eeb2-42b6-97ba-1497de3030bc.png">
 
Focus state:
<img width="305" alt="image"
src="https://user-images.githubusercontent.com/7453188/227027678-c0f0242b-f374-4997-9584-92cddb06fdf6.png">

## Test Plan 
- Manual testing (regular/imported/Ledger accounts)

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
N/A
