+++ apps/wallet/src/ui/app/components/menu/content/Account.tsx
@@ -6,11 +6,11 @@ import { ChevronDown16, Copy16 } from '@mysten/icons';
 import { formatAddress } from '@mysten/sui.js';
 import { cx } from 'class-variance-authority';
 
+import { AccountBadge } from '../../AccountBadge';
 import { AccountActions } from './AccountActions';
-import { AccountType } from '_src/background/keyring/Account';
+import { type AccountType } from '_src/background/keyring/Account';
 import { useCopyToClipboard } from '_src/ui/app/hooks/useCopyToClipboard';
 import { Heading } from '_src/ui/app/shared/heading';
-import { Text } from '_src/ui/app/shared/text';
 
 export type AccountProps = {
     address: string;
@@ -73,33 +73,4 @@ export function Account({ address, accountType }: AccountProps) {
             )}
         </Disclosure>
     );
-}
-
-type AccountBadgeProps = {
-    accountType: AccountType;
-};
-
-function AccountBadge({ accountType }: AccountBadgeProps) {
-    let badgeText: string | null = null;
-    switch (accountType) {
-        case AccountType.LEDGER:
-            badgeText = 'Ledger';
-            break;
-        case AccountType.IMPORTED:
-            badgeText = 'Imported';
-            break;
-        case AccountType.DERIVED:
-            badgeText = null;
-            break;
-        default:
-            throw new Error(`Encountered unknown account type ${accountType}`);
-    }
-
-    return badgeText ? (
-        <div className="bg-gray-40 rounded-2xl border border-solid border-gray-45 py-1 px-1.5">
-            <Text variant="captionSmallExtra" color="steel-dark">
-                {badgeText}
-            </Text>
-        </div>
-    ) : null;
 }
